Reorganize components in places that make more sense
