+++ src/shared/components/Profile/ProfileAvatar.tsx
